requirements.txt
absl-py==0.9.0
amqp==2.5.2
anyjson==0.3.3
appdirs==1.4.3
asgiref==3.2.3
astor==0.8.1
billiard==3.6.3.0
cachetools==4.0.0
celery==4.4.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
distlib==0.3.0
Django==3.0.3
django-celery==3.3.1
django-celery-results==1.2.0
django-cors-headers==3.2.1
django-redis==4.11.0
django-redis-sessions==0.6.1
djangorestframework==3.11.0
efficientnet==1.1.0
filelock==3.0.12
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
imageio==2.8.0
importlib-metadata==1.5.0
itsdangerous==1.1.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kombu==4.6.8
Markdown==3.1.1
matplotlib==3.1.3
networkx==2.4
nose==1.3.7
numpy==1.18.1
oauthlib==3.1.0
opencv-contrib-python==4.2.0.32
opt-einsum==3.1.0
Pillow==7.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
redis==3.4.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scipy==1.4.1
six==1.12.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.8
vine==1.3.0
virtualenv==20.0.7
Werkzeug==0.16.1
wrapt==1.11.2
zipp==2.2.0
python-dotenv==0.12.0
watchdog==0.10.2
argh==0.26.2